feat: Add CSS to homepage
This commit completely changes the homepage style.
Previously, it was plain text. Now, there's a box
wrapped in the middle of the page.

The recipes are also in buttons, wrapped in a
flex display, with added 'flex-wrap' and 'justify-content'.

I also fixed the h1 font sizes, and removed the
random 'mark' on one of the recipes.

Finally, the html, body, and all * selectors are reset.
That means we'll have to add padding on all
recipe pages in the future.